Fixed the way the client works so it has Connect and Disconnect.
Created a model for connections with the server and client.

Edited SteeringModel so it checks if the client is supposed to be connected before telling it to send data.

Created FlightButtons viewModel that controls the buttons of "Connect" and "Settings".

Fixed FlightBoard.cs so it displays new points (hopefully it would work).

General bug fixes.